Major Login authentication update
Fixed login authentication.
Fixed styling for showing/hiding edit/delete buttons
Added nav bar for login/logout functionality
Updated templates to be able to use a base.html template with jinja2
Updated authentication method for checking if the user is logged in or not using a GET request and a newly added route.