Update start-project.py
Migrating google authentication callback to the main project.

Added a Log method for printing information to the server to make things easier to debug

Added Get_User_Info() and Add_User() methods for querying the db for the User information, and adding users to the db.

Implemented check logic to determine if a user needs to be added or not.